+++ code/client/rule/vnc/process/process_darwin.go
@@ -1,0 +1,11 @@
+package process
+
+import (
+	"image"
+	"time"
+)
+
+// CreateWorker create worker process
+func CreateWorker(name, confDir string, showCursor bool) (*Process, error) {
+	return nil, ErrNotSupported
+}
